requirements.txt
python>=3.10
qiskit>=2.0
qiskit-algorithms>=0.4.0
qiskit-aer>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
ipywidgets>=8.0.0
pandas>=2.0.0
pyyaml>=6.0.0
tqdm>=4.65.0
pytest>=7.4.0
